Bug - Promotion Rule dialog tooltips are incorrect -
all tooltips say 'Please provide a name with no spaces.  This name is also used in any API requests where this field is allocated.'

so for example 'Enabled' tooltip is 'Please provide a name with no spaces.  This name is also used in any API requests where this field is allocated.'