Added logic to be able to mark field types as being Editable, Displayable.  This now also applies to fusion attributes, where you can mark certain fields as editable.  More testing needed around this.
Added Web Analytics tables, procedure, view, and extended the CompanyContext to account for this procedure call.  Instead of sending web analytic over to table storage, it now saves the record in the client database.  This will help in reporting.  Also now track views for dashboards.

Updated a number of nuget packages across the solution.