GOV-16437 GOV-17724 Fix issue with not updated breadcrumbs on several pages
The issue was caused by incorrect checking if we need to apply interceptor. Instead of checking if request is Query, we checked if request method is GET. This made issue with ignorance of caching on those queries which are POST-requests. So, we tried to sent request, which will be cancelled because of page change