Stop using d360.core.constants in d360.model.JobDbContextCreator
The reason for this move is the chain of callings:
igx.functions.consumption (Azure Functions) -> JobDbContextCreator -> constants -> ConfigurationManager.AppSettings

The last step in this chain (ConfigurationManager.AppSettings) does not work when called inside the Azure Function and always returns NULL.
In addition to this, `d360.core.constants` is a static class with only static fields. It leads to System.ArgumentNullException thrown from `public static readonly int COMPANY_BULK_LOAD_MAX_ROWS = int.Parse(ConfigurationManager.AppSettings["BulkLoadMaxRows"]);` every time we call to any field from the `constants` class. It works that way because all the fields in a static class will be calculated when an instance of the class is created.