GOV-16437 Fix caching to correctly handle POST queries
Some POST requests are marked as queries, so they are cached. In order to get cache key, we need not only to use url, but body too.

Note that JSON.stringify isn't guaranteed to be stable, but is good enough for us. In case if we will have problems, you can use json-stable-stringify