GOV-17232 Fix workflow editor stuck on loading screen
IsLoading status is not correctly handled in workflow-diagram.component.ts. There are several lines where isLoading changed, so we had race condition. Depending on the speed of server APIs we can either observe bugs or not.
After introducing reuse.interceptor.ts our HTTP requests are stored in-memory cache. This made this race-condition bug to be reproducible in 100% of cases when we reopen the diagram editor.

Simply using isLoadingCounter++ instead of isLoading=true solves this situation.