GOV-17290 Fix issue that after "unselect all" previous checkboxes have wrong style
This issue originally reported for Admin/Tags page in comment for a bug https://infogix.atlassian.net/browse/GOV-17290?focusedCommentId=338509. I found that we have same issue for Configuration/Diagram Assets/Connector Labels page, so fixed that too.

Issue happens because we are incorrectly working with primeng library. Primeng library expects us to pass new array when there are changes in contents of array. Instead we simply mutate our DOM-tree. This naive approach doesn't scales well and leads to bugs. Instead of this, we will create a copy of mutated array, so primeng will be able to rerender tree.