GOV-17722 Fix "Number of rows" is set to the wrong category.
The issue happens, because we are trying to set numberOfRows depending on breadcrumbs & so we use old breadcrumbs.
In order to fix this issue, we maintain 'pendingChanges' and apply pendingChanges when proper breadcrumbs are loaded.

This approach is incompatible with old schema ({"Reference Lists":{"Reference Lists":25},"Technical Assets":100}). So we migrate to new more explicit schema ({"Reference Lists":{"Reference Lists":25},"Technical Assets":{"Main": 100}})