GOV-18630 Fix showing form badges on Scoring Definitions
This issue happens because we are incorrectly rate-limit updating visibility of required/invalid badges.
We use debounce function, which doesn't guarantee that updating will ever happen. Instead, when a large amount of attempts to update hapens, debounce function prevents updating at all. So, we should use throttle function which guarantees that update will be called at least once in between of 200ms.