GOV-18067 Use new api/v2/survey/types on UI (temporary)
Unfortunately, this API doesn't fits well for UI.
It produces a paged response, but it doesn't provides ability to set order direction & filters.
Thus we request 100000 items and hope that this amount of items covers all surveys.
It's really bad, but otherwise we need to rewrite old internal api or rewrite new api, which is out of scope of this task