Fix for GOV-19018 and GOV-19020
I'm not a big fan of fixing multiple bugs in one PR, but here it is reasonable.
So, these changes will switch from HTTP 500 to HTTP 401 for case when user is not authorized to delete particular tag. In addition to it we will return better error message when invalig UID has been sent