GOV-17794 Make kebab menu sticky
Please, note that in order to make kebab menu sticky, we need to make sure that table cell has same background as table row.
Otherwise sticky table cell is transparent and this makes a mess on UI. This is why we use `background: inherit`
This is why I have changed EnvironmentController to produce not rgba(x,x,x,0.2), but rgb(y,y,y) - because inheriting transparent colors works badly.