GOV-19281 Wrong count value after overriding responsibility(cherry-pick)
- Changed sql query for selecting count of assets for specified ResponsibilityTypeUid and AssetTypeUid.

Short explanation how select count query works:
It is selecting rows without duplicates.
In case when there is only one row with AssetID = 0 in dbo.ResponsibilityDetail table for ResponsibilityTypeId = 11111 query will include this row in summary count but in case when dbo.ResponsibilityDetail table has this row and row (or rows) with AssetID !=0 query will ignore row with AssetID = 0 and will add only rows with AssetID != 0.