GOV-19498 Tags: fix handling of windows resize
Previously, the issue was that tags doesn't increase width when we increase width of window.
Unfortunattely, my previous fix was incorrect and lead to new issue. New issue is that sometimes tag overlaps other items after resize.

This happens, because when we calculate width of collapsed tag - it doesn't overflows parent and we consider this as "tag is small enough" → we allow it to take fullwidth → it overflows in some cases.

So new algo is a bit simpler:
1. Always force tag to take full width
2. If it overflows parent, force it to become smaller
3. If it doesn't overflow, do nothing