GOV-19902 Fix advanced filtering for fields from relationship
The root issue is that in the code, we had a forceful replacement of data in fieldType. But this was invalid because part of the code relays that data is original while other parts of the code relays that data is replaced.

This issue happened after the performance fix by introducing TempTableFieldToken, but it doesn't matter - it's just a matter of time before something breaks the original code.

The fix is simple: whenever we need not unoriginal data, use unoriginal data directly and don't mess with original data.